requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.3
django-crispy-forms==1.6.1
django-filter==1.0.4
django-formtools==2.0
django-guardian==1.4.9
django-reversion==2.0.10
djangorestframework==3.6.4
future==0.16.0
httplib2==0.10.3
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
mysqlclient==1.3.10
olefile==0.44
Pillow==4.2.1
pytz==2017.2
requests==2.18.4
six==1.10.0
uritemplate==3.0.0
urllib3==1.22
XlsxWriter==0.9.8
xlwt==1.3.0
